+++ app/oz/built_ins/index.js
@@ -1,0 +1,11 @@
+import Number from "./number";
+import Float from "./float";
+import Record from "./record";
+
+export const builtIns = {
+  Number,
+  Float,
+  Record,
+};
+
+export const allBuiltInTypes = Object.keys(builtIns);
